Fix total word count bug
Total word count for text streams containing multiple headings would be
equal to *only* the word count of the final section, because the word
count of prior sections would be thrown away with each new section. I
have fixed this by retaining the word count of prior sections before
saving and by adding a wrinkle to the stats tests.

It's possible we would prefer to cover this just be getting the word
count from the stored stat blocks in the long run.